Response to comments
