+++ packages/neo-one-monitor/src/NodeMonitor.js
@@ -7,13 +7,12 @@ import mount from 'koa-mount';
 import perfHooks from 'perf_hooks';
 import prom from 'prom-client';
 
-import type { LogLevel, Monitor } from './types';
+import type { LogLevel, Monitor, MetricConstruct } from './types';
 import MonitorBase, {
   type CounterMetric,
   type GaugeMetric,
   type HistogramMetric,
   type Logger,
-  type MetricConstruct,
   type MetricsFactory,
   type SummaryMetric,
   type Tracer,
